+++ pyrgg/test.py
@@ -45,7 +45,7 @@ True
 >>> result = input_filter({"file_name": "test2","vertices": 23,"max_weight": 2,"min_weight": 80,"min_edge": 23,"max_edge": 1,"sign": True,"output_format": 1, "direct": False,"self_loop": True,"multigraph":False,"number_of_files":100,"engine":1})
 >>> result == {'min_weight': 2, 'vertices': 23, 'file_name': 'test2', 'max_edge': 23, 'min_edge': 1, 'max_weight': 80, 'output_format': 1, 'sign': True, "direct": False,"self_loop": True,"multigraph":False,"number_of_files":100,"engine":1}
 True
->>> logger('test',100,50,1000,10,1,0,0,1,20,1,'2min')
+>>> logger('test',100,50,1000,10,1,0,0,1,20,1,1,'2min')
 >>> file=open('logfile.log','r')
 >>> print("\n".join(file.read().splitlines()[1:-1]))
 Filename : test
@@ -60,6 +60,7 @@ Self Loop : True
 Weighted : True
 Max Weight : 20
 Min Weight : 1
+Engine : 1 (pyrgg)
 Elapsed Time : 2min
 >>> convert_bytes(200)
 '200.0 bytes'