+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Generation engine menu
 ### Changed
+- `engine` parameter added to `logger` function
 - `MENU_ITEMS1` parameter changed to `MENU_ITEMS`
 - `MENU_ITEMS2` parameter changed to `PYRGG_ENGINE_PARAMS`
 - `_update_using_first_menu` function changed to `_update_using_menu`